AQUISIÇÃO-TRATAMENTO-E-VISUALIZAÇÃO-DE-DADOS/server/requirements.txt
altair==5.1.2
ansi2html==1.8.0
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.2
Pillow==10.1.0
plotly==5.17.0
protobuf==4.25.0
pyarrow==14.0.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
streamlit==1.28.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
Werkzeug==2.2.3
zipp==3.17.0
